requirements.txt
alabaster==0.7.12
auto-sklearn==0.5.1
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
ConfigSpace==0.4.10
Cython==0.29.7
docutils==0.14
idna==2.8
imagesize==1.1.0
Jinja2==2.10.1
joblib==0.13.2
liac-arff==2.4.0
lockfile==0.12.2
MarkupSafe==1.1.1
nose==1.3.7
numpy==1.16.3
packaging==19.0
pandas==0.24.2
psutil==5.6.1
Pygments==2.3.1
pynisher==0.5.0
pyparsing==2.4.0
pyrfr==0.7.4
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
scikit-learn==0.19.2
scipy==1.2.1
six==1.12.0
smac==0.8.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
swiglpk==4.65.0
typing==3.6.6
urllib3==1.24.2
xgboost==0.82
